various fixes
* corrections for simple data types in ROS message types according to ROS specification
* moving message generation classes to Unity3D project
* resourceFileUri fix
* Prismatic Joint translation fix
* updating dependencies

+++ Libraries/RosBridgeClient/Messages/Standard/Float64.cs
@@ -21,11 +21,11 @@ namespace RosSharp.RosBridgeClient.Messages.Standard
     {
         [JsonIgnore]
         public const string RosMessageName = "std_msgs/Float64";
-        public float data;
+        public double data;
 
         public Float64()
         {
-            data = 0.0f;
+            data = 0.0;
         }
     }
 }
